user-service/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
SQLAlchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic==2.6.1
pydantic-settings==2.1.0
greenlet==3.0.3
email-validator==2.1.0
bcrypt==4.0.1